requirements.txt
cycler==0.10.0
gitdb2==2.0.5
GitPython==2.1.11
kiwisolver==1.0.1
lizard==1.16.3
matplotlib==3.0.3
mysql==0.0.2
mysqlclient==1.4.2.post1
numpy==1.16.2
Pillow==6.0.0
PyDriller==1.7
PyMySQL==0.9.3
pyparsing==2.3.1
PyQt5==5.12.1
PyQt5-sip==4.19.15
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
six==1.12.0
smmap2==2.0.5
SQLAlchemy==1.3.1
typing==3.6.6
